refactor: organize integration tests into separate files
- Created separate test files for each integration controller: articles, comments, topics, and API.
Created a `setupTestDB` file to centralize database connection setup and reseeding logic.
